Draft of refactor
