+++ fast64_internal/f3d/f3d_writer.py
@@ -1,24 +1,14 @@
 from typing import Union
 import functools
-import bpy, bmesh, mathutils, os, re, copy, math
-from math import pi, ceil
-from io import BytesIO
+import bpy, mathutils, os, re, copy, math
+from math import ceil
 from bpy.utils import register_class, unregister_class
 
 from .f3d_enums import *
 from .f3d_constants import *
-from .f3d_material import (
-    all_combiner_uses,
-    getMaterialScrollDimensions,
-    getTmemWordUsage,
-    getTmemMax,
-    texBitSizeF3D,
-    texFormatOf,
-    TextureProperty,
-    F3DMaterialProperty,
-)
+from .f3d_material import all_combiner_uses, getMaterialScrollDimensions
+from .f3d_texture_writer import MultitexManager
 from .f3d_gbi import *
-from .f3d_gbi import _DPLoadTextureBlock
 
 from ..utility import *
 
@@ -217,237 +207,6 @@ def findUVBounds(polygon, uv_data):
             minUV[i] = uv[i] if minUV[i] is None else min(minUV[i], uv[i])
             maxUV[i] = uv[i] if maxUV[i] is None else max(maxUV[i], uv[i])
     return minUV, maxUV
-
-
-class TileLoad:
-    def __init__(self, material, fMaterial, texDimensions):
-        self.sl = self.tl = 1000000  # above any actual value
-        self.sh = self.th = -1  # below any actual value
-
-        self.texFormat = fMaterial.largeTexFmt
-        self.is4bit = texBitSizeInt[self.texFormat] == 4
-        self.tmemWordsAvail = fMaterial.largeTexWords
-        self.texDimensions = texDimensions
-        self.materialName = material.name
-        self.isPointSampled = isTexturePointSampled(material)
-        self.largeEdges = material.f3d_mat.large_edges
-
-        self.faces = []
-        self.offsets = []
-
-    def getLow(self, value, field):
-        value = int(math.floor(value))
-        if self.largeEdges == "Clamp":
-            value = min(max(value, 0), self.texDimensions[field] - 1)
-        if self.is4bit and field == 0:
-            # Must start on an even texel (round down)
-            value &= ~1
-        return value
-
-    def getHigh(self, value, field):
-        value = int(math.ceil(value)) - (1 if self.isPointSampled else 0)
-        if self.largeEdges == "Clamp":
-            value = min(max(value, 0), self.texDimensions[field] - 1)
-        if self.is4bit and field == 0:
-            # Must end on an odd texel (round up)
-            value |= 1
-        return value
-
-    def fixRegion(self, sl, sh, tl, th):
-        assert sl <= sh and tl <= th
-        soffset = int(math.floor(sl / self.texDimensions[0])) * self.texDimensions[0]
-        toffset = int(math.floor(tl / self.texDimensions[1])) * self.texDimensions[1]
-        sl -= soffset
-        sh -= soffset
-        tl -= toffset
-        th -= toffset
-        assert 0 <= sl < self.texDimensions[0] and 0 <= tl < self.texDimensions[1]
-        ret = True
-        if sh >= 1024 or th >= 1024:
-            ret = False
-        if sh >= self.texDimensions[0]:
-            # Load wraps in S. Load must start a multiple of a TMEM word from
-            # the end of the texture, in order for the second load (beginning of
-            # image) to start at a whole word.
-            texelsPerWord = 64 // texBitSizeInt[self.texFormat]
-            if texelsPerWord > self.texDimensions[0]:
-                raise PluginError(
-                    f"In large texture material {self.materialName}:"
-                    + f" large texture must be at least {texelsPerWord} wide."
-                )
-            sl -= self.texDimensions[0]
-            sl = int(math.floor(sl / texelsPerWord)) * texelsPerWord
-            sl += self.texDimensions[0]
-        if th >= self.texDimensions[1]:
-            # Load wraps in T. Load must start a multiple of 2 texture rows from
-            # the end of the texture, in order for the second load to have the
-            # same odd/even row parity as the first (because texels are
-            # interleaved in TMEM every other row).
-            tl -= self.texDimensions[1]
-            tl = int(math.floor(tl / 2.0)) * 2
-            tl += self.texDimensions[1]
-        tmemUsage = getTmemWordUsage(self.texFormat, sh - sl + 1, th - tl + 1)
-        if tmemUsage > self.tmemWordsAvail:
-            ret = False
-        return ret, sl, sh, tl, th, soffset, toffset
-
-    def initWithFace(self, obj, face):
-        uv_data = obj.data.uv_layers["UVMap"].data
-        faceUVs = [UVtoSTLarge(obj, loopIndex, uv_data, self.texDimensions) for loopIndex in face.loops]
-        if len(faceUVs) == 0:
-            return True
-
-        for point in faceUVs:
-            self.sl = min(self.sl, self.getLow(point[0], 0))
-            self.sh = max(self.sh, self.getHigh(point[0], 0))
-            self.tl = min(self.tl, self.getLow(point[1], 1))
-            self.th = max(self.th, self.getHigh(point[1], 1))
-
-        ret, self.sl, self.sh, self.tl, self.th, soffset, toffset = self.fixRegion(self.sl, self.sh, self.tl, self.th)
-        if not ret:
-            if self.sh >= 1024 or self.th >= 1024:
-                raise PluginError(
-                    f"Large texture material {self.materialName} has a face that needs"
-                    + f" to cover texels {self.sl}-{self.sh} x {self.tl}-{self.th}"
-                    + f" (image dims are {self.texDimensions}), but image space"
-                    + f" only goes up to 1024 so this cannot be represented."
-                )
-            else:
-                raise PluginError(
-                    f"Large texture material {self.materialName} has a face that needs"
-                    + f" to cover texels {self.sl}-{self.sh} x {self.tl}-{self.th}"
-                    + f" ({self.sh-self.sl+1} x {self.th-self.tl+1} texels) "
-                    + f"in format {self.texFormat}, which can't fit in TMEM."
-                )
-        self.faces.append(face)
-        self.offsets.append((soffset, toffset))
-
-    def trySubsume(self, other):
-        # Could do fancier logic checking across borders, for example if we have
-        # one loading 60-68 (size 64) and another 0-8, that could be merged to
-        # one load 60-72. But this is likely to be uncommon and won't be generated
-        # by the operator.
-        new_sl = min(self.sl, other.sl)
-        new_sh = max(self.sh, other.sh)
-        new_tl = min(self.tl, other.tl)
-        new_th = max(self.th, other.th)
-        ret, new_sl, new_sh, new_tl, new_th, soffset, toffset = self.fixRegion(new_sl, new_sh, new_tl, new_th)
-        if not ret:
-            return False
-        self.sl, self.sh, self.tl, self.th = new_sl, new_sh, new_tl, new_th
-        self.faces.extend(other.faces)
-        self.offsets.extend(other.offsets)
-        return True
-
-
-def maybeSaveSingleLargeTextureSetup(
-    i: int,
-    fMaterial: FMaterial,
-    fModel: FModel,
-    fImage: FImage,
-    gfxOut: GfxList,
-    texProp: TextureProperty,
-    texDimensions: tuple[int, int],
-    tileSettings: TileLoad,
-    curImgSet: Union[None, int],
-    curTileLines: list[int],
-):
-    if fMaterial.isTexLarge[i]:
-        wrapS = tileSettings.sh >= texDimensions[0]
-        wrapT = tileSettings.th >= texDimensions[1]
-        assert 0 <= tileSettings.sl < texDimensions[0]
-        assert 0 <= tileSettings.tl < texDimensions[1]
-        siz = texBitSizeF3D[texProp.tex_format]
-        line = getTileLine(fImage, tileSettings.sl, tileSettings.sh, siz, fModel.f3d)
-        tmem = fMaterial.largeTexAddr[i]
-        # print(
-        #     f"Tile: {tileSettings.sl}-{tileSettings.sh} x {tileSettings.tl}-{tileSettings.th} "
-        #     + f"tmem {tmem} line {line}"
-        # )
-        if wrapS or wrapT:
-            fmt = texFormatOf[texProp.tex_format]
-            texelsPerWord = 64 // texBitSizeInt[texProp.tex_format]
-            wid = texDimensions[0]
-            is4bit = siz == "G_IM_SIZ_4b"
-            if is4bit:
-                siz = "G_IM_SIZ_8b"
-                wid >>= 1
-                assert (tileSettings.sl & 1) == 0
-                assert (tileSettings.sh & 1) == 1
-            # TL, TH is always * 4 because tile values are 10.2 fixed.
-            # SL, SH is * 2 for 4 bit and * 4 otherwise, because actually loading
-            # 8 bit pairs of texels. Also written using f3d.G_TEXTURE_IMAGE_FRAC.
-            sm = 2 if is4bit else 4
-            nocm = ["G_TX_WRAP", "G_TX_NOMIRROR"]
-            if curImgSet != i:
-                gfxOut.commands.append(DPSetTextureImage(fmt, siz, wid, fImage))
-
-            def loadOneOrTwoS(tmemBase, tidxBase, TL, TH):
-                if line != curTileLines[tidxBase]:
-                    gfxOut.commands.append(DPSetTile(fmt, siz, line, tmemBase, tidxBase, 0, nocm, 0, 0, nocm, 0, 0))
-                    curTileLines[tidxBase] = line
-                if wrapS:
-                    # Break up at the wrap boundary into two tile loads.
-                    # The first load must occupy a whole number of lines.
-                    assert (texDimensions[0] - tileSettings.sl) % texelsPerWord == 0
-                    sLineOfs = (texDimensions[0] - tileSettings.sl) // texelsPerWord
-                    # print(f"-- Wrap at S={texDimensions[0]}, offset {sLineOfs}")
-                    gfxOut.commands.append(
-                        DPLoadTile(tidxBase, tileSettings.sl * sm, TL * 4, (texDimensions[0] - 1) * sm, TH * 4)
-                    )
-                    gfxOut.commands.append(
-                        DPSetTile(fmt, siz, line, tmemBase + sLineOfs, tidxBase - 1, 0, nocm, 0, 0, nocm, 0, 0)
-                    )
-                    curTileLines[tidxBase - 1] = -1
-                    gfxOut.commands.append(
-                        DPLoadTile(tidxBase - 1, 0, TL * 4, (tileSettings.sh - texDimensions[0]) * sm, TH * 4)
-                    )
-                else:
-                    gfxOut.commands.append(
-                        DPLoadTile(tidxBase, tileSettings.sl * sm, TL * 4, tileSettings.sh * sm, TH * 4)
-                    )
-
-            if wrapT:
-                # Break up at the wrap boundary into two loads.
-                # The first load must be even in size (even number of texture rows).
-                assert (texDimensions[1] - tileSettings.tl) % 2 == 0
-                tLineOfs = line * (texDimensions[1] - tileSettings.tl)
-                # print(f"-- Wrap at T={texDimensions[1]}, offset {tLineOfs}")
-                loadOneOrTwoS(tmem, 7, tileSettings.tl, texDimensions[1] - 1)
-                loadOneOrTwoS(tmem + tLineOfs, 5, 0, tileSettings.th - texDimensions[1])
-            else:
-                loadOneOrTwoS(tmem, 7, tileSettings.tl, tileSettings.th)
-            if fMaterial.isTexLarge[i ^ 1]:
-                # May reuse any of the above tiles for the other large texture.
-                gfxOut.commands.append(DPTileSync())
-        else:
-            saveTextureLoadOnly(
-                fImage,
-                gfxOut,
-                texProp,
-                tileSettings,
-                7 - i,
-                tmem,
-                fModel.f3d,
-                curImgSet == i,
-                line == curTileLines[7 - i],
-            )
-            curTileLines[7 - i] = line
-        curImgSet = i
-        saveTextureTile(
-            fImage,
-            fMaterial,
-            gfxOut,
-            texProp,
-            tileSettings,
-            i,
-            tmem,
-            fMaterial.texPaletteIndex[i],
-            fModel.f3d,
-            line == curTileLines[i],
-        )
-        curTileLines[i] = line
-    return curImgSet
 
 
 def saveMeshWithLargeTexturesByFaces(
@@ -1608,65 +1367,7 @@ def saveOrGetF3DMaterial(material, fModel, obj, drawLayer, convertTextureData):
         )
 
     useDict = all_combiner_uses(f3dMat)
-
-    # Get texture info, needed for othermode; also check some texture props
-    err0, info0 = getTexInfoFromMat(0, f3dMat)
-    err1, info1 = getTexInfoFromMat(1, f3dMat)
-    if err0 is not None:
-        raise PluginError(f"In {material.name} tex0: {err0}")
-    if err1 is not None:
-        raise PluginError(f"In {material.name} tex1: {err1}")
-    (useTex0, isTex0Ref, isTex0CI, tex0Fmt, pal0Fmt, imageDims0, tex0Tmem) = info0
-    (useTex1, isTex1Ref, isTex1CI, tex1Fmt, pal1Fmt, imageDims1, tex1Tmem) = info1
-    pal0, pal0Len, im0Use, tex0Flipbook = getTexInfoAdvanced(
-        0, material, fMaterial, fModel, useTex0, isTex0Ref, isTex0CI, tex0Fmt, pal0Fmt
-    )
-    pal1, pal1Len, im1Use, tex1Flipbook = getTexInfoAdvanced(
-        1, material, fMaterial, fModel, useTex1, isTex1Ref, isTex1CI, tex1Fmt, pal1Fmt
-    )
-
-    isCI = (useTex0 and isTex0CI) or (useTex1 and isTex1CI)
-    isPal0Ref = isTex0Ref and tex0Flipbook is None
-    isPal1Ref = isTex1Ref and tex1Flipbook is None
-
-    if useTex0 and useTex1:
-        if isTex0CI != isTex1CI:
-            raise PluginError(
-                "In material "
-                + material.name
-                + ": N64 does not support CI + non-CI texture. "
-                + "Must be both CI or neither CI."
-            )
-        if (
-            isTex0Ref
-            and isTex1Ref
-            and f3dMat.tex0.tex_reference == f3dMat.tex1.tex_reference
-            and f3dMat.tex0.tex_reference_size != f3dMat.tex1.tex_reference_size
-        ):
-            raise PluginError(
-                "In material " + material.name + ": Two textures with the same reference must have the same size."
-            )
-        if isCI:
-            if pal0Fmt != pal1Fmt:
-                raise PluginError(
-                    "In material "
-                    + material.name
-                    + ": Both CI textures must use the same palette format (usually RGBA16)."
-                )
-            if (
-                isTex0Ref
-                and isTex1Ref
-                and f3dMat.tex0.pal_reference == f3dMat.tex1.pal_reference
-                and f3dMat.tex0.pal_reference_size != f3dMat.tex1.pal_reference_size
-            ):
-                raise PluginError(
-                    "In material "
-                    + material.name
-                    + ": Two textures with the same palette reference must have the same palette size."
-                )
-
-    palFormat = pal0Fmt if useTex0 else pal1Fmt
-    g_tt = "G_TT_NONE" if not isCI else ("G_TT_" + palFormat)
+    multitexManager = MultitexManager(material, fMaterial, fModel)
 
     # Set othermode
     if drawLayer is not None:
@@ -1680,7 +1381,7 @@ def saveOrGetF3DMaterial(material, fModel, obj, drawLayer, convertTextureData):
     else:
         saveGeoModeDefinition(fMaterial, f3dMat.rdp_settings, defaults, fModel.matWriteMethod)
     saveOtherModeHDefinition(
-        fMaterial, f3dMat.rdp_settings, g_tt, defaults, fModel.f3d._HW_VERSION_1, fModel.matWriteMethod
+        fMaterial, f3dMat.rdp_settings, multitexManager.getTT(), defaults, fModel.f3d._HW_VERSION_1, fModel.matWriteMethod
     )
     saveOtherModeLDefinition(fMaterial, f3dMat.rdp_settings, defaults, defaultRM, fModel.matWriteMethod)
     saveOtherDefinition(fMaterial, f3dMat, defaults)
@@ -1695,317 +1396,10 @@ def saveOrGetF3DMaterial(material, fModel, obj, drawLayer, convertTextureData):
     else:
         fMaterial.material.commands.append(SPTexture(s, t, 0, fModel.f3d.G_TX_RENDERTILE, 1))
 
-    # Determine how to arrange / load palette entries into upper half of tmem
-    tex0PaletteIndex = 0
-    tex1PaletteIndex = 0
-    loadPal0 = False
-    loadPal1 = False
-    pal0Addr = 0
-    pal1Addr = 0
-    pal0Use = im0Use
-    pal1Use = im1Use
-    if isCI:
-        assert useTex0 or useTex1
-        if not useTex1:
-            loadPal0 = True
-        elif not useTex0:
-            loadPal1 = True
-        elif not convertTextureData:
-            if tex0Fmt == "CI8" or tex1Fmt == "CI8":
-                raise PluginError(
-                    "In material "
-                    + material.name
-                    + ": When using export as PNGs mode, can't have multitexture with one or more CI8 textures."
-                    + " Only single CI texture or two CI4 textures."
-                )
-            loadPal0 = loadPal1 = True
-            tex1PaletteIndex = 1
-            pal1Addr = 16
-        else:  # Two CI textures, normal mode
-            if tex0Fmt == "CI8" and tex1Fmt == "CI8":
-                if (pal0 is None) != (pal1 is None):
-                    raise PluginError(
-                        "In material "
-                        + material.name
-                        + ": can't have two CI8 textures where only one is a non-flipbook reference; "
-                        + "no way to assign the palette."
-                    )
-                loadPal0 = True
-                if pal0 is None:
-                    if f3dMat.tex0.pal_reference != f3dMat.tex1.pal_reference:
-                        raise PluginError(
-                            "In material "
-                            + material.name
-                            + ": can't have two CI8 textures with different palette references."
-                        )
-                else:
-                    pal0 = mergePalettes(pal0, pal1)
-                    pal0Len = len(pal0)
-                    if pal0Len > 256:
-                        raise PluginError(
-                            "In material "
-                            + material.name
-                            + ": the two CI textures together contain a total of "
-                            + str(pal0Len)
-                            + " colors, which can't fit in a CI8 palette (256)."
-                        )
-                    # im0Use remains what it was; the CIs in im0 are the same as they
-                    # would be if im0 was alone. But im1 and pal0 depend on both.
-                    im1Use = pal0Use = im0Use + im1Use
-            elif tex0Fmt != tex1Fmt:  # One CI8, one CI4
-                ci8Pal, ci4Pal = (pal0, pal1) if tex0Fmt == "CI8" else (pal1, pal0)
-                ci8PalLen, ci4PalLen = (pal0Len, pal1Len) if tex0Fmt == "CI8" else (pal1Len, pal0Len)
-                if pal0 is None or pal1 is None:
-                    if ci8PalLen > 256 - 16:
-                        raise PluginError(
-                            "In material "
-                            + material.name
-                            + ": the CI8 texture has over 240 colors, which can't fit together with the CI4 palette."
-                        )
-                    loadPal0 = loadPal1 = True
-                    if tex0Fmt == "CI8":
-                        tex1PaletteIndex = 15
-                        pal1Addr = 240
-                    else:
-                        tex0PaletteIndex = 15
-                        pal0Addr = 240
-                else:
-                    # CI4 indices in palette 0, CI8 indices start from palette 0
-                    loadPal0 = True
-                    pal0 = mergePalettes(ci4Pal, ci8Pal)
-                    pal0Len = len(pal0)
-                    if pal0Len > 256:
-                        raise PluginError(
-                            "In material "
-                            + material.name
-                            + ": the two CI textures together contain a total of "
-                            + str(pal0Len)
-                            + " colors, which can't fit in a CI8 palette (256)."
-                            + " The CI8 texture must contain up to 240 unique colors,"
-                            + " plus the same up to 16 colors used in the CI4 texture."
-                        )
-                    # The use for the CI4 texture remains what it was; its CIs are the
-                    # same as if it was alone. But both the palette and the CI8 CIs are affected.
-                    pal0Use = im0Use + im1Use
-                    if tex0Fmt == "CI8":
-                        im0Use = pal0Use
-                    else:
-                        im1Use = pal0Use
-            else:  # both CI4 textures
-                if pal0 is None and pal1 is None and f3dMat.tex0.pal_reference == f3dMat.tex1.pal_reference:
-                    loadPal0 = True
-                elif pal0 is None or pal1 is None:
-                    loadPal0 = loadPal1 = True
-                    tex1PaletteIndex = 1
-                    pal1Addr = 16
-                else:
-                    loadPal0 = True
-                    tempPal = mergePalettes(pal0, pal1)
-                    tempPalLen = len(tempPal)
-                    assert tempPalLen <= 32
-                    if tempPalLen <= 16:
-                        # Share palette 0
-                        pal0 = tempPal
-                        pal0Len = tempPalLen
-                        # im0Use remains what it was; the CIs in im0 are the same as they
-                        # would be if im0 was alone. But im1 and pal0 depend on both.
-                        im1Use = pal0Use = im0Use + im1Use
-                    else:
-                        # Load one palette across 0-1. Put the longer in slot 0
-                        if pal0Len >= pal1Len:
-                            while len(pal0) < 16:
-                                pal0.append(0)
-                            pal0.extend(pal1)
-                            pal0Len = len(pal0)
-                            tex1PaletteIndex = 1
-                        else:
-                            while len(pal1) < 16:
-                                pal1.append(0)
-                            pal0 = pal1 + pal0
-                            pal0Len = len(pal0)
-                            tex0PaletteIndex = 1
-                        # The up-to-32 entries in pal0 depend on both images. But the
-                        # CIs in both im0 and im1 are the same as if there was no shared palette.
-                        pal0Use = im0Use + im1Use
-    fMaterial.texPaletteIndex = [tex0PaletteIndex, tex1PaletteIndex]
-    pal0BaseName = getPaletteName(useTex0, isPal0Ref, f3dMat.tex0, tex0Flipbook)
-    pal1BaseName = getPaletteName(useTex1, isPal1Ref, f3dMat.tex1, tex1Flipbook)
-    if isCI and useTex0 and useTex1 and not loadPal1:
-        pal0BaseName = pal0BaseName + "_x_" + pal1BaseName
-        pal1 = pal0
-
-    # Assign TMEM addresses
-    sameTextures = (
-        useTex0
-        and useTex1
-        and (
-            (not isTex0Ref and not isTex1Ref and f3dMat.tex0.tex == f3dMat.tex1.tex)
-            or (isTex0Ref and isTex1Ref and f3dMat.tex0.tex_reference == f3dMat.tex1.tex_reference)
-        )
-    )
-    useLargeTextures = material.mat_ver > 3 and f3dMat.use_large_textures
-    tmemSize = 256 if isCI else 512
-    doTex0Load = doTex0Tile = doTex1Load = doTex1Tile = True
-    tex1Addr = None  # must be set whenever tex 1 used (and loaded or tiled)
-    tmemOccupied = texDimensions = None  # must be set on all codepaths
-    if sameTextures:
-        assert tex0Tmem == tex1Tmem
-        tmemOccupied = tex0Tmem
-        doTex1Load = False
-        tex1Addr = 0
-        texDimensions = imageDims0
-        fMaterial.largeTexFmt = tex0Fmt
-    elif not useLargeTextures or tex0Tmem + tex1Tmem <= tmemSize:
-        tex1Addr = tex0Tmem
-        tmemOccupied = tex0Tmem + tex1Tmem
-        if not useTex0 and not useTex1:
-            texDimensions = [32, 32]
-            fMaterial.largeTexFmt = "RGBA16"
-        elif not useTex1 or f3dMat.uv_basis == "TEXEL0":
-            texDimensions = imageDims0
-            fMaterial.largeTexFmt = tex0Fmt
-        else:
-            texDimensions = imageDims1
-            fMaterial.largeTexFmt = tex1Fmt
-    else:  # useLargeTextures
-        if useTex0 and useTex1:
-            tmemOccupied = tmemSize
-            # TODO: Could change this in the future to do the face tile assigments
-            # first, to see how large a tile the large texture(s) needed, instead
-            # of arbitrarily assigning half of TMEM to each of the two textures.
-            if tex0Tmem <= tmemSize // 2:
-                # Tex 0 normal, tex 1 large
-                texDimensions = imageDims1
-                fMaterial.largeTexFmt = tex1Fmt
-                fMaterial.isTexLarge[1] = True
-                fMaterial.largeTexAddr[1] = tex0Tmem
-                fMaterial.largeTexWords = tmemSize - tex0Tmem
-                doTex1Load = doTex1Tile = False
-            elif tex1Tmem <= tmemSize // 2:
-                # Tex 0 large, tex 1 normal
-                texDimensions = imageDims0
-                fMaterial.largeTexFmt = tex0Fmt
-                fMaterial.isTexLarge[0] = True
-                fMaterial.largeTexAddr[0] = 0
-                fMaterial.largeTexWords = tmemSize - tex1Tmem
-                doTex0Load = doTex0Tile = False
-                tex1Addr = tmemSize - tex1Tmem
-            else:
-                # Both textures large
-                raise PluginError(
-                    'Error in "' + material.name + '": Multitexture with two large textures is not currently supported.'
-                )
-                # Limited cases of 2x large textures could be supported in the
-                # future. However, these cases are either of questionable
-                # utility or have substantial restrictions. Most cases could be
-                # premixed into one texture, or would run out of UV space for
-                # tiling (1024x1024 in the space of whichever texture had
-                # smaller pixels), or one of the textures could be non-large.
-                if f3dMat.uv_basis == "TEXEL0":
-                    texDimensions = imageDims0
-                    fMaterial.largeTexFmt = tex0Fmt
-                else:
-                    texDimensions = imageDims1
-                    fMaterial.largeTexFmt = tex1Fmt
-                fMaterial.isTexLarge[0] = True
-                fMaterial.isTexLarge[1] = True
-                fMaterial.largeTexAddr[0] = 0
-                fMaterial.largeTexAddr[1] = tmemSize // 2
-                fMaterial.largeTexWords = tmemSize // 2
-                doTex0Load = doTex0Tile = doTex1Load = doTex1Tile = False
-        elif useTex0:
-            texDimensions = imageDims0
-            fMaterial.largeTexFmt = tex0Fmt
-            fMaterial.isTexLarge[0] = True
-            fMaterial.largeTexAddr[0] = 0
-            fMaterial.largeTexWords = tmemSize
-            doTex0Load = doTex0Tile = False
-            tmemOccupied = tmemSize
-        elif useTex1:
-            tex1Addr = 0
-            texDimensions = imageDims1
-            fMaterial.largeTexFmt = tex1Fmt
-            fMaterial.isTexLarge[1] = True
-            fMaterial.largeTexAddr[1] = 0
-            fMaterial.largeTexWords = tmemSize
-            doTex1Load = doTex1Tile = False
-            tmemOccupied = tmemSize
-    if tmemOccupied > tmemSize:
-        if sameTextures and useLargeTextures:
-            raise PluginError(
-                'Error in "'
-                + material.name
-                + '": Using the same texture for Tex0 and Tex1 is not compatible with large textures.'
-            )
-        elif not bpy.context.scene.ignoreTextureRestrictions:
-            raise PluginError(
-                'Error in "'
-                + material.name
-                + '": Textures are too big. Max TMEM size is 4k '
-                + "bytes, ex. 2 32x32 RGBA 16 bit textures.\nNote that texture width will be internally padded to 64 bit boundaries."
-            )
-
-    # Get texture and palette definitions
-    fImage0 = fImage1 = fPalette0 = fPalette1 = None
-    if useTex0:
-        imageKey0, fImage0 = saveOrGetTextureDefinition(
-            fMaterial, fModel, f3dMat.tex0, im0Use, fMaterial.isTexLarge[0]
-        )
-        fMaterial.imageKey[0] = imageKey0
-    if loadPal0:
-        paletteKey0, fPalette0 = saveOrGetPaletteDefinition(fMaterial, fModel, f3dMat.tex0, isPal0Ref, pal0Use, pal0BaseName, pal0Len)
-    if useTex1:
-        imageKey1, fImage1 = saveOrGetTextureDefinition(
-            fMaterial, fModel, f3dMat.tex1, im1Use, fMaterial.isTexLarge[1]
-        )
-        fMaterial.imageKey[1] = imageKey1
-    if loadPal1:
-        paletteKey1, fPalette1 = saveOrGetPaletteDefinition(fMaterial, fModel, f3dMat.tex1, isPal1Ref, pal1Use, pal1BaseName, pal1Len)
-
-    # Write DL entries to load textures and palettes
-    loadGfx = fMaterial.material
-    if loadPal0:
-        savePaletteLoad(loadGfx, fPalette0, pal0Fmt, pal0Addr, pal0Len, 5, fModel.f3d)
-    if useTex0 and doTex0Load:
-        saveTextureLoadOnly(fImage0, loadGfx, f3dMat.tex0, None, 7, 0, fModel.f3d)
-    if useTex0 and doTex0Tile:
-        saveTextureTile(fImage0, fMaterial, loadGfx, f3dMat.tex0, None, 0, 0, tex0PaletteIndex, fModel.f3d)
-    if loadPal1:
-        savePaletteLoad(loadGfx, fPalette1, pal1Fmt, pal1Addr, pal1Len, 4, fModel.f3d)
-    if useTex1 and doTex1Load:
-        saveTextureLoadOnly(fImage1, loadGfx, f3dMat.tex1, None, 6, tex1Addr, fModel.f3d)
-    if useTex1 and doTex1Tile:
-        saveTextureTile(fImage1, fMaterial, loadGfx, f3dMat.tex1, None, 1, tex1Addr, tex1PaletteIndex, fModel.f3d)
-
-    # Write texture and palette data, unless exporting textures as PNGs.
-    if convertTextureData:
-        if loadPal0 and not isPal0Ref:
-            writePaletteData(fPalette0, pal0)
-        if useTex0:
-            if isTex0Ref:
-                if isCI:
-                    fModel.writeTexRefCITextures(tex0Flipbook, fMaterial, im0Use, pal0, tex0Fmt, pal0Fmt)
-                else:
-                    fModel.writeTexRefNonCITextures(tex0Flipbook, tex0Fmt)
-            else:
-                if isCI:
-                    writeCITextureData(f3dMat.tex0.tex, fImage0, pal0, pal0Fmt, tex0Fmt)
-                else:
-                    writeNonCITextureData(f3dMat.tex0.tex, fImage0, tex0Fmt)
-        if loadPal1 and not isPal1Ref:
-            writePaletteData(fPalette1, pal1)
-        if useTex1:
-            if isTex1Ref:
-                if isCI:
-                    fModel.writeTexRefCITextures(tex1Flipbook, fMaterial, im1Use, pal1, tex1Fmt, pal1Fmt)
-                else:
-                    fModel.writeTexRefNonCITextures(tex1Flipbook, tex1Fmt)
-            else:
-                if isCI:
-                    writeCITextureData(f3dMat.tex1.tex, fImage1, pal1, pal1Fmt, tex1Fmt)
-                else:
-                    writeNonCITextureData(f3dMat.tex1.tex, fImage1, tex1Fmt)
-
+    # Write textures
+    multitexManager.writeAll(material, fMaterial, fModel, convertTextureData)
+    
+    # Write colors
     nodes = material.node_tree.nodes
     if useDict["Primitive"] and f3dMat.set_prim:
         color = exportColor(f3dMat.prim_color[0:3]) + [scaleToU8(f3dMat.prim_color[3])]
@@ -2078,6 +1472,7 @@ def saveOrGetF3DMaterial(material, fModel, obj, drawLayer, convertTextureData):
     else:
         fMaterial.revert = None
 
+    texDimensions = multitexManager.getTexDimensions()
     materialKey = (
         material,
         (drawLayer if f3dMat.rdp_settings.set_rendermode else None),
@@ -2086,711 +1481,6 @@ def saveOrGetF3DMaterial(material, fModel, obj, drawLayer, convertTextureData):
     fModel.materials[materialKey] = (fMaterial, texDimensions)
 
     return fMaterial, texDimensions
-
-
-# Functions for texture and palette definitions
-
-
-def getTextureNamesFromBasename(
-    baseName: str,
-    texOrPalFormat: str,
-    parent: Union[FModel, FTexRect],
-    isPalette: bool
-):
-    suffix = getTextureSuffixFromFormat(texOrPalFormat)
-    imageName = parent.name + "_" + baseName + "_"
-    if isPalette:
-        imageName += "pal_"
-    imageName += suffix
-    imageName = checkDuplicateTextureName(parent, toAlnum(imageName))
-    filename = baseName + "." + suffix + (".pal" if isPalette else ".inc.c")
-    return imageName, filename
-
-
-def getImageName(image: bpy.types.Image):
-    if image is None:
-        raise PluginError("No image set in material!")
-    elif image.filepath == "":
-        return image.name
-    else:
-        return getNameFromPath(image.filepath, True)
-
-
-def getTextureNamesFromImage(
-    image: bpy.types.Image,
-    texFormat: str,
-    parent: Union[FModel, FTexRect]
-):
-    return getTextureNamesFromBasename(getImageName(image), texFormat, parent, False)
-
-
-def getTextureNamesFromProp(
-    texProp: TextureProperty,
-    parent: Union[FModel, FTexRect]
-):
-    if texProp.use_tex_reference:
-        raise PluginError("Internal error, invalid use of getTextureNamesFromProp")
-    return getTextureNamesFromImage(texProp.tex, texProp.tex_format, parent)
-
-
-def getPaletteName(
-    useTex: bool,
-    isPalRef: bool,
-    texProp: TextureProperty,
-    flipbook: Union[None, "TextureFlipbook"],
-):
-    if not useTex or isPalRef:
-        return None
-    if flipbook is not None:
-        return flipbook.name
-    return getImageName(texProp.tex)
-
-
-def checkDuplicateTextureName(parent: Union[FModel, FTexRect], name):
-    names = []
-    for info, texture in parent.textures.items():
-        names.append(texture.name)
-    while name in names:
-        name = name + "_copy"
-    return name
-
-
-def saveOrGetPaletteDefinition(
-    fMaterial: FMaterial,
-    parent: Union[FModel, FTexRect],
-    texProp: TextureProperty,
-    isPalRef: bool,
-    images: list[bpy.types.Image],
-    palBaseName: str,
-    palLen: int,
-) -> tuple[FPaletteKey, FImage]:
-
-    texFmt = texProp.tex_format
-    palFmt = texProp.ci_format
-    palFormat = texFormatOf[palFmt]
-    paletteKey = FPaletteKey(palFmt, images)
-
-    if isPalRef:
-        fPalette = FImage(texProp.pal_reference, None, None, 1, palLen, None)
-        return paletteKey, fPalette
-
-    # If palette already loaded, return that data.
-    fPalette = parent.getTextureAndHandleShared(paletteKey)
-    if fPalette is not None:
-        # print(f"Palette already exists")
-        return paletteKey, fPalette
-
-    paletteName, filename = getTextureNamesFromBasename(palBaseName, palFmt, parent, True)
-    fPalette = FImage(paletteName, palFormat, "G_IM_SIZ_16b", 1, palLen, filename)
-
-    parent.addTexture(paletteKey, fPalette, fMaterial)
-    return paletteKey, fPalette
-
-
-def saveOrGetTextureDefinition(
-    fMaterial: FMaterial,
-    parent: Union[FModel, FTexRect],
-    texProp: TextureProperty,
-    images: list[bpy.types.Image],
-    isLarge: bool,
-) -> tuple[FImageKey, FImage]:
-
-    image = texProp.tex
-    texFmt = texProp.tex_format
-    texFormat = texFormatOf[texFmt]
-    bitSize = texBitSizeF3D[texFmt]
-    imageKey = getImageKey(texProp, images)
-
-    if texProp.use_tex_reference:
-        width, height = texProp.tex_reference_size
-        fImage = FImage(texProp.tex_reference, None, None, width, height, None)
-        return imageKey, fImage
-
-    # If image already loaded, return that data.
-    fImage = parent.getTextureAndHandleShared(imageKey)
-    if fImage is not None:
-        # print(f"Image already exists")
-        return imageKey, fImage
-
-    imageName, filename = getTextureNamesFromProp(texProp, parent)
-    fImage = FImage(imageName, texFormat, bitSize, image.size[0], image.size[1], filename)
-    fImage.isLargeTexture = isLarge
-
-    parent.addTexture(imageKey, fImage, fMaterial)
-    return imageKey, fImage
-
-
-def getTexInfoFromMat(
-    index: int,
-    f3dMat: F3DMaterialProperty,
-):
-    texProp = getattr(f3dMat, "tex" + str(index))
-
-    useDict = all_combiner_uses(f3dMat)
-    if not useDict["Texture " + str(index)]:
-        return None, (False, False, False, "", "", (0, 0), 0)
-
-    return getTexInfoFromProp(texProp)
-
-
-def getTexInfoFromProp(texProp: TextureProperty):
-    if not texProp.tex_set:
-        return None, (False, False, False, "", "", (0, 0), 0)
-
-    tex = texProp.tex
-    isTexRef = texProp.use_tex_reference
-    texFormat = texProp.tex_format
-    isCITexture = texFormat[:2] == "CI"
-    palFormat = texProp.ci_format if isCITexture else ""
-
-    if tex is not None and (tex.size[0] == 0 or tex.size[1] == 0):
-        return f"Image {tex.name} has 0 size; may have been deleted/moved.", None
-
-    if not isTexRef:
-        if tex is None:
-            return f"No texture is selected.", None
-        elif len(tex.pixels) == 0:
-            return f"Image {tex.name} is missing on disk.", None
-
-    if isTexRef:
-        width, height = texProp.tex_reference_size
-    else:
-        width, height = tex.size
-
-    tmemSize = getTmemWordUsage(texFormat, width, height)
-
-    if width > 1024 or height > 1024:
-        return f"Image size (even large textures) limited to 1024 in each dimension.", None
-
-    if texBitSizeInt[texFormat] == 4 and (width & 1) != 0:
-        return f"A 4-bit image must have a width which is even.", None
-
-    info = (True, isTexRef, isCITexture, texFormat, palFormat, (width, height), tmemSize)
-    return None, info
-
-
-def getTexInfoAdvanced(
-    index: int,
-    material: bpy.types.Material,
-    fMaterial: FMaterial,
-    fModel: FModel,
-    useTex: bool,
-    isTexRef: bool,
-    isCITexture: bool,
-    texFormat: str,
-    palFormat: str,
-):
-    if not useTex:
-        return None, 0, [], None
-    
-    f3dMat = material.f3d_mat
-    texProp = getattr(f3dMat, "tex" + str(index))
-
-    pal = None
-    palLen = 0
-    if isCITexture:
-        imUse, flipbook, pal = fModel.processTexRefCITextures(fMaterial, material, index)
-        if isTexRef:
-            if flipbook is not None:
-                palLen = len(pal)
-            else:
-                palLen = texProp.pal_reference_size
-        else:
-            assert flipbook is None
-            pal = getColorsUsedInImage(texProp.tex, palFormat)
-            palLen = len(pal)
-        if palLen > (16 if texFormat == "CI4" else 256):
-            raise PluginError(
-                f"Error in {material.name}: texture {index}"
-                + (" (all flipbook textures)" if flipbook is not None else "")
-                + f" uses too many unique colors to fit in format {texFormat}."
-            )
-    else:
-        imUse, flipbook = fModel.processTexRefNonCITextures(fMaterial, material, index)
-
-    return pal, palLen, imUse, flipbook
-
-
-# Functions for writing texture and palette DLs
-
-
-def getTileSizeSettings(texProp: TextureProperty, tileSettings: Union[None, TileLoad], f3d: F3D):
-    if tileSettings is not None:
-        SL = tileSettings.sl
-        TL = tileSettings.tl
-        SH = tileSettings.sh
-        TH = tileSettings.th
-    else:
-        SL = texProp.S.low
-        TL = texProp.T.low
-        SH = texProp.S.high
-        TH = texProp.T.high
-    sl = int(SL * (2**f3d.G_TEXTURE_IMAGE_FRAC))
-    tl = int(TL * (2**f3d.G_TEXTURE_IMAGE_FRAC))
-    sh = int(SH * (2**f3d.G_TEXTURE_IMAGE_FRAC))
-    th = int(TH * (2**f3d.G_TEXTURE_IMAGE_FRAC))
-    return SL, TL, SH, TH, sl, tl, sh, th
-
-
-def getTileLine(fImage: FImage, SL: int, SH: int, siz: str, f3d: F3D):
-    width = int(SH - SL + 1) if fImage.isLargeTexture else int(fImage.width)
-    if siz == "G_IM_SIZ_4b":
-        line = (((width + 1) >> 1) + 7) >> 3
-    else:
-        # Note that _LINE_BYTES and _TILE_BYTES variables are the same.
-        line = int((width * f3d.G_IM_SIZ_VARS[siz + "_LINE_BYTES"]) + 7) >> 3
-    return line
-
-
-def canUseLoadBlock(fImage: FImage, tex_format: str, f3d: F3D):
-    if fImage.isLargeTexture:
-        return False
-    width, height = fImage.width, fImage.height
-    texelsPerWord = 64 // texBitSizeInt[tex_format]
-    if width % texelsPerWord != 0:
-        return False
-    wordsperrow = width // texelsPerWord
-    dxt = ((1 << f3d.G_TX_DXT_FRAC) + wordsperrow - 1) // wordsperrow
-    error = (dxt * wordsperrow) - (1 << f3d.G_TX_DXT_FRAC)
-    assert error >= 0
-    if error == 0:
-        return True
-    rowsWhenCorruptionHappens = (dxt + error - 1) // error
-    return height < rowsWhenCorruptionHappens
-
-
-def saveTextureLoadOnly(
-    fImage: FImage,
-    gfxOut: GfxList,
-    texProp: TextureProperty,
-    tileSettings: Union[None, TileLoad],
-    loadtile: int,
-    tmem: int,
-    f3d: F3D,
-    omitSetTextureImage=False,
-    omitSetTile=False,
-):
-    fmt = texFormatOf[texProp.tex_format]
-    siz = texBitSizeF3D[texProp.tex_format]
-    nocm = ["G_TX_WRAP", "G_TX_NOMIRROR"]
-    SL, TL, SH, TH, sl, tl, sh, th = getTileSizeSettings(texProp, tileSettings, f3d)
-
-    # LoadTile will pad rows to 64 bit word alignment, while
-    # LoadBlock assumes this is already done.
-    useLoadBlock = canUseLoadBlock(fImage, texProp.tex_format, f3d)
-    line = 0 if useLoadBlock else getTileLine(fImage, SL, SH, siz, f3d)
-    wid = 1 if useLoadBlock else fImage.width
-
-    if siz == "G_IM_SIZ_4b":
-        if useLoadBlock:
-            dxs = (((fImage.width) * (fImage.height) + 3) >> 2) - 1
-            dxt = f3d.CALC_DXT_4b(fImage.width)
-            siz = "G_IM_SIZ_16b"
-            loadCommand = DPLoadBlock(loadtile, 0, 0, dxs, dxt)
-        else:
-            sl2 = int(SL * (2 ** (f3d.G_TEXTURE_IMAGE_FRAC - 1)))
-            sh2 = int(SH * (2 ** (f3d.G_TEXTURE_IMAGE_FRAC - 1)))
-            siz = "G_IM_SIZ_8b"
-            wid >>= 1
-            loadCommand = DPLoadTile(loadtile, sl2, tl, sh2, th)
-    else:
-        if useLoadBlock:
-            dxs = (
-                ((fImage.width) * (fImage.height) + f3d.G_IM_SIZ_VARS[siz + "_INCR"])
-                >> f3d.G_IM_SIZ_VARS[siz + "_SHIFT"]
-            ) - 1
-            dxt = f3d.CALC_DXT(fImage.width, f3d.G_IM_SIZ_VARS[siz + "_BYTES"])
-            siz += "_LOAD_BLOCK"
-            loadCommand = DPLoadBlock(loadtile, 0, 0, dxs, dxt)
-        else:
-            loadCommand = DPLoadTile(loadtile, sl, tl, sh, th)
-
-    if not omitSetTextureImage:
-        gfxOut.commands.append(DPSetTextureImage(fmt, siz, wid, fImage))
-    if not omitSetTile:
-        gfxOut.commands.append(DPSetTile(fmt, siz, line, tmem, loadtile, 0, nocm, 0, 0, nocm, 0, 0))
-    gfxOut.commands.append(loadCommand)
-
-
-def saveTextureTile(
-    fImage: FImage,
-    fMaterial: FMaterial,
-    gfxOut: GfxList,
-    texProp: TextureProperty,
-    tileSettings,
-    rendertile: int,
-    tmem: int,
-    pal: int,
-    f3d: F3D,
-    omitSetTile=False,
-):
-    if tileSettings is not None:
-        clamp_S = True
-        clamp_T = True
-        mirror_S = False
-        mirror_T = False
-        mask_S = 0
-        mask_T = 0
-        shift_S = 0
-        shift_T = 0
-    else:
-        clamp_S = texProp.S.clamp
-        clamp_T = texProp.T.clamp
-        mirror_S = texProp.S.mirror
-        mirror_T = texProp.T.mirror
-        mask_S = texProp.S.mask
-        mask_T = texProp.T.mask
-        shift_S = texProp.S.shift
-        shift_T = texProp.T.shift
-    cms = [("G_TX_CLAMP" if clamp_S else "G_TX_WRAP"), ("G_TX_MIRROR" if mirror_S else "G_TX_NOMIRROR")]
-    cmt = [("G_TX_CLAMP" if clamp_T else "G_TX_WRAP"), ("G_TX_MIRROR" if mirror_T else "G_TX_NOMIRROR")]
-    masks = mask_S
-    maskt = mask_T
-    shifts = shift_S if shift_S >= 0 else (shift_S + 16)
-    shiftt = shift_T if shift_T >= 0 else (shift_T + 16)
-    fmt = texFormatOf[texProp.tex_format]
-    siz = texBitSizeF3D[texProp.tex_format]
-    SL, _, SH, _, sl, tl, sh, th = getTileSizeSettings(texProp, tileSettings, f3d)
-    line = getTileLine(fImage, SL, SH, siz, f3d)
-
-    tileCommand = DPSetTile(fmt, siz, line, tmem, rendertile, pal, cmt, maskt, shiftt, cms, masks, shifts)
-    tileSizeCommand = DPSetTileSize(rendertile, sl, tl, sh, th)
-    tileSizeCommand.tags |= GfxTag.TileScroll0 if rendertile == 0 else GfxTag.TileScroll1
-    tileSizeCommand.fMaterial = fMaterial
-    if not omitSetTile:
-        gfxOut.commands.append(tileCommand)
-    gfxOut.commands.append(tileSizeCommand)
-
-    # hasattr check for FTexRect
-    if hasattr(fMaterial, "tileSizeCommands"):
-        fMaterial.tileSizeCommands[rendertile] = tileSizeCommand
-
-
-# palAddr is the address within the second half of tmem (0-255), normally 16*palette num
-# palLen is the number of colors
-def savePaletteLoad(
-    gfxOut: GfxList,
-    fPalette: FImage,
-    palFormat: str,
-    palAddr: int,
-    palLen: int,
-    loadtile: int,
-    f3d: F3D,
-):
-    assert 0 <= palAddr < 256 and (palAddr & 0xF) == 0
-    palFmt = texFormatOf[palFormat]
-    nocm = ["G_TX_WRAP", "G_TX_NOMIRROR"]
-
-    if not f3d._HW_VERSION_1:
-        gfxOut.commands.extend(
-            [
-                DPSetTextureImage(palFmt, "G_IM_SIZ_16b", 1, fPalette),
-                DPSetTile("0", "0", 0, 256 + palAddr, loadtile, 0, nocm, 0, 0, nocm, 0, 0),
-                DPLoadTLUTCmd(loadtile, palLen - 1),
-            ]
-        )
-    else:
-        gfxOut.commands.extend(
-            [
-                _DPLoadTextureBlock(
-                    fPalette,
-                    256 + palAddr,
-                    palFmt,
-                    "G_IM_SIZ_16b",
-                    4 * palLen,
-                    1,
-                    0,
-                    nocm,
-                    nocm,
-                    0,
-                    0,
-                    0,
-                    0,
-                )
-            ]
-        )
-
-
-# Functions for converting and writing texture and palette data
-
-
-def extractConvertCIPixel(image, pixels, i, j, palFormat):
-    color = [1, 1, 1, 1]
-    for field in range(image.channels):
-        color[field] = pixels[(j * image.size[0] + i) * image.channels + field]
-    if palFormat == "RGBA16":
-        pixelColor = getRGBA16Tuple(color)
-    elif palFormat == "IA16":
-        pixelColor = getIA16Tuple(color)
-    else:
-        raise PluginError("Internal error, palette format is " + palFormat)
-    return pixelColor
-
-
-def getColorsUsedInImage(image, palFormat):
-    palette = []
-    # N64 is -Y, Blender is +Y
-    pixels = image.pixels[:]
-    for j in reversed(range(image.size[1])):
-        for i in range(image.size[0]):
-            pixelColor = extractConvertCIPixel(image, pixels, i, j, palFormat)
-            if pixelColor not in palette:
-                palette.append(pixelColor)
-    return palette
-
-
-def mergePalettes(pal0, pal1):
-    palette = [c for c in pal0]
-    for c in pal1:
-        if c not in palette:
-            palette.append(c)
-    return palette
-
-
-def getColorIndicesOfTexture(image, palette, palFormat):
-    texture = []
-    # N64 is -Y, Blender is +Y
-    pixels = image.pixels[:]
-    for j in reversed(range(image.size[1])):
-        for i in range(image.size[0]):
-            pixelColor = extractConvertCIPixel(image, pixels, i, j, palFormat)
-            if pixelColor not in palette:
-                raise PluginError(f"Bug: {image.name} palette len {len(palette)} missing CI")
-            texture.append(palette.index(pixelColor))
-    return texture
-
-
-def compactNibbleArray(texture, width, height):
-    nibbleData = bytearray(0)
-    dataSize = int(width * height / 2)
-
-    nibbleData = [((texture[i * 2] & 0xF) << 4) | (texture[i * 2 + 1] & 0xF) for i in range(dataSize)]
-
-    if (width * height) % 2 == 1:
-        nibbleData.append((texture[-1] & 0xF) << 4)
-
-    return bytearray(nibbleData)
-
-
-def writePaletteData(fPalette: FImage, palette: list[int]):
-    if fPalette.converted:
-        return
-    for color in palette:
-        fPalette.data.extend(color.to_bytes(2, "big"))
-    fPalette.converted = True
-
-
-def writeCITextureData(
-    image: bpy.types.Image,
-    fImage: FImage,
-    palette: list[int],
-    palFmt: str,
-    texFmt: str,
-):
-    if fImage.converted:
-        return
-
-    texture = getColorIndicesOfTexture(image, palette, palFmt)
-
-    if texFmt == "CI4":
-        fImage.data = compactNibbleArray(texture, image.size[0], image.size[1])
-    else:
-        fImage.data = bytearray(texture)
-    fImage.converted = True
-
-
-def writeNonCITextureData(image: bpy.types.Image, fImage: FImage, texFmt: str):
-    if fImage.converted:
-        return
-    fmt = texFormatOf[texFmt]
-    bitSize = texBitSizeF3D[texFmt]
-
-    pixels = image.pixels[:]
-    if fmt == "G_IM_FMT_RGBA":
-        if bitSize == "G_IM_SIZ_16b":
-            fImage.data = bytearray(
-                [
-                    byteVal
-                    for doubleByte in [
-                        (
-                            (
-                                ((int(round(pixels[(j * image.size[0] + i) * image.channels + 0] * 0x1F)) & 0x1F) << 3)
-                                | (
-                                    (int(round(pixels[(j * image.size[0] + i) * image.channels + 1] * 0x1F)) & 0x1F)
-                                    >> 2
-                                )
-                            ),
-                            (
-                                ((int(round(pixels[(j * image.size[0] + i) * image.channels + 1] * 0x1F)) & 0x03) << 6)
-                                | (
-                                    (int(round(pixels[(j * image.size[0] + i) * image.channels + 2] * 0x1F)) & 0x1F)
-                                    << 1
-                                )
-                                | (1 if pixels[(j * image.size[0] + i) * image.channels + 3] > 0.5 else 0)
-                            ),
-                        )
-                        for j in reversed(range(image.size[1]))
-                        for i in range(image.size[0])
-                    ]
-                    for byteVal in doubleByte
-                ]
-            )
-        elif bitSize == "G_IM_SIZ_32b":
-            fImage.data = bytearray(
-                [
-                    int(round(pixels[(j * image.size[0] + i) * image.channels + field] * 0xFF)) & 0xFF
-                    for j in reversed(range(image.size[1]))
-                    for i in range(image.size[0])
-                    for field in range(image.channels)
-                ]
-            )
-        else:
-            raise PluginError("Invalid combo: " + fmt + ", " + bitSize)
-
-    elif fmt == "G_IM_FMT_YUV":
-        raise PluginError("YUV not yet implemented.")
-        if bitSize == "G_IM_SIZ_16b":
-            pass
-        else:
-            raise PluginError("Invalid combo: " + fmt + ", " + bitSize)
-
-    elif fmt == "G_IM_FMT_CI":
-        raise PluginError("CI not yet implemented.")
-
-    elif fmt == "G_IM_FMT_IA":
-        if bitSize == "G_IM_SIZ_4b":
-            fImage.data = bytearray(
-                [
-                    (
-                        (
-                            int(
-                                round(
-                                    colorToLuminance(
-                                        pixels[
-                                            (j * image.size[0] + i)
-                                            * image.channels : (j * image.size[0] + i)
-                                            * image.channels
-                                            + 3
-                                        ]
-                                    )
-                                    * 0x7
-                                )
-                            )
-                            & 0x7
-                        )
-                        << 1
-                    )
-                    | (1 if pixels[(j * image.size[0] + i) * image.channels + 3] > 0.5 else 0)
-                    for j in reversed(range(image.size[1]))
-                    for i in range(image.size[0])
-                ]
-            )
-        elif bitSize == "G_IM_SIZ_8b":
-            fImage.data = bytearray(
-                [
-                    (
-                        (
-                            int(
-                                round(
-                                    colorToLuminance(
-                                        pixels[
-                                            (j * image.size[0] + i)
-                                            * image.channels : (j * image.size[0] + i)
-                                            * image.channels
-                                            + 3
-                                        ]
-                                    )
-                                    * 0xF
-                                )
-                            )
-                            & 0xF
-                        )
-                        << 4
-                    )
-                    | (int(round(pixels[(j * image.size[0] + i) * image.channels + 3] * 0xF)) & 0xF)
-                    for j in reversed(range(image.size[1]))
-                    for i in range(image.size[0])
-                ]
-            )
-        elif bitSize == "G_IM_SIZ_16b":
-            fImage.data = bytearray(
-                [
-                    byteVal
-                    for doubleByte in [
-                        (
-                            int(
-                                round(
-                                    colorToLuminance(
-                                        pixels[
-                                            (j * image.size[0] + i)
-                                            * image.channels : (j * image.size[0] + i)
-                                            * image.channels
-                                            + 3
-                                        ]
-                                    )
-                                    * 0xFF
-                                )
-                            )
-                            & 0xFF,
-                            int(round(pixels[(j * image.size[0] + i) * image.channels + 3] * 0xFF)) & 0xFF,
-                        )
-                        for j in reversed(range(image.size[1]))
-                        for i in range(image.size[0])
-                    ]
-                    for byteVal in doubleByte
-                ]
-            )
-        else:
-            raise PluginError("Invalid combo: " + fmt + ", " + bitSize)
-    elif fmt == "G_IM_FMT_I":
-        if bitSize == "G_IM_SIZ_4b":
-            fImage.data = bytearray(
-                [
-                    int(
-                        round(
-                            colorToLuminance(
-                                pixels[
-                                    (j * image.size[0] + i) * image.channels : (j * image.size[0] + i) * image.channels
-                                    + 3
-                                ]
-                            )
-                            * 0xF
-                        )
-                    )
-                    & 0xF
-                    for j in reversed(range(image.size[1]))
-                    for i in range(image.size[0])
-                ]
-            )
-        elif bitSize == "G_IM_SIZ_8b":
-            fImage.data = bytearray(
-                [
-                    int(
-                        round(
-                            colorToLuminance(
-                                pixels[
-                                    (j * image.size[0] + i) * image.channels : (j * image.size[0] + i) * image.channels
-                                    + 3
-                                ]
-                            )
-                            * 0xFF
-                        )
-                    )
-                    & 0xFF
-                    for j in reversed(range(image.size[1]))
-                    for i in range(image.size[0])
-                ]
-            )
-        else:
-            raise PluginError("Invalid combo: " + fmt + ", " + bitSize)
-    else:
-        raise PluginError("Invalid image format " + fmt)
-
-    # We stored 4bit values in byte arrays, now to convert
-    if bitSize == "G_IM_SIZ_4b":
-        fImage.data = compactNibbleArray(fImage.data, image.size[0], image.size[1])
-
-    fImage.converted = True
 
 
 def saveLightsDefinition(fModel, fMaterial, material, lightsName):
